# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/lab/moods": {
      "get": {
        "operationId": "LabController_moods_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All moods with their Spanish aliases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mood"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Available moods",
        "tags": [
          "Lab"
        ]
      }
    },
    "/v1/lab/moods/{mood}": {
      "get": {
        "description": "Picks one of the 10 best matches at random (plus 3 alternatives). `focused` and `hungover` are alcohol-free; `party` needs alcohol access.",
        "operationId": "LabController_byMood_v1",
        "parameters": [
          {
            "name": "mood",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Mood key or Spanish alias",
              "example": "guayabo"
            },
            "description": "Mood key or alias"
          }
        ],
        "responses": {
          "200": {
            "description": "The recommendation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoodRecommendation"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: role, ownership or age restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "A drink for your mood",
        "tags": [
          "Lab"
        ]
      }
    },
    "/v1/lab/pantry": {
      "get": {
        "description": "What you can make with what you have, what you almost can, and which single purchase unlocks the most recipes. Understands Spanish names (ron, limón, hierbabuena…); assumes ice and water.",
        "operationId": "LabController_pantry_v1",
        "parameters": [
          {
            "name": "have",
            "required": false,
            "in": "query",
            "schema": {
              "description": "Comma separated ingredients, English or Spanish",
              "example": "ron,limón,azúcar,hierbabuena,soda",
              "default": "",
              "type": "string"
            },
            "description": "Comma separated ingredients, English or Spanish"
          },
          {
            "name": "maxMissing",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "description": "Max missing ingredients for a drink to appear in `almost`",
              "type": "integer",
              "minimum": 0,
              "maximum": 3
            },
            "description": "Max missing ingredients for a drink to appear in `almost`"
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "description": "Max items per list",
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Max items per list"
          },
          {
            "name": "alcoholic",
            "required": false,
            "in": "query",
            "schema": {
              "description": "true: only alcoholic; false: only alcohol-free; omit: any",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "true: only alcoholic; false: only alcohol-free; omit: any"
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantrySuggestions"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Smart pantry",
        "tags": [
          "Lab"
        ]
      }
    },
...
  },
  "info": {
    "title": "api-drinks",
    "description": "\nAPI de bebidas y cocteles sobre TheCocktailDB: despensa inteligente, cocteles por estado de ánimo,\nADN de sabor, gemelos y carta inteligente para bares.\n\n**Versión**: todas las rutas van bajo `/v1`, salvo `/health`, `/health/ready` y `/`.\n\n**Errores**: todos tienen la forma `{ statusCode, code, message, error, details? }`. El frontend decide por `code`\n(estable) y muestra su propio texto; `message` es solo una pista en inglés.\n\n**Autenticación**\n- Apps propias: `Authorization: Bearer <accessToken>` (ver /v1/auth/login). En navegador, el refresh token\n  viaja en una cookie httpOnly: haz las peticiones de /v1/auth con credentials: 'include'.\n- Integraciones de terceros: `X-API-Key: <key>` (ver /v1/me/api-keys). Cuentan contra la cuota diaria del plan\n  y responden con los headers `X-RateLimit-Limit`, `X-RateLimit-Remaining` y `X-RateLimit-Reset`.\n\n**Fuente de datos**: recetas e imágenes de [TheCocktailDB](https://www.thecocktaildb.com/). Todo frontend que\nmuestre bebidas debe mostrar \"Datos e imágenes: TheCocktailDB\" con un enlace a su sitio (condición de su uso).\n\n**Verificación de edad**: las bebidas con alcohol solo se muestran a usuarios autenticados mayores de 18 años.\n",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "example": 403
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PLAN_LIMIT",
              "RATE_LIMITED",
              "UNAVAILABLE",
              "VALIDATION_FAILED",
              "ROUTE_NOT_FOUND",
              "PAYLOAD_TOO_LARGE",
              "INTERNAL_ERROR",
              "AUTH_REQUIRED",
              "INVALID_CREDENTIALS",
              "INVALID_ACCESS_TOKEN",
              "INVALID_REFRESH_TOKEN",
              "INVALID_API_KEY",
              "API_KEY_NOT_ALLOWED",
              "AMBIGUOUS_CREDENTIALS",
              "ORIGIN_NOT_ALLOWED",
              "ROLE_REQUIRED",
              "AGE_RESTRICTED",
              "LOGIN_LOCKED",
              "API_QUOTA_EXCEEDED",
              "EMAIL_TAKEN",
              "INVALID_EMAIL",
              "WEAK_PASSWORD",
              "INVALID_BIRTH_DATE",
              "ADMIN_SELF_DEMOTION",
              "ADMIN_SELF_DELETION",
              "WRONG_PASSWORD",
              "INVALID_RESET_TOKEN",
              "INVALID_NAME",
              "USER_NOT_FOUND",
              "DRINK_NOT_FOUND",
              "MOOD_NOT_FOUND",
              "CATALOG_UNAVAILABLE",
              "NO_TASTE_YET",
              "TASTE_LINK_NOT_FOUND",
              "OWN_TASTE_LINK",
              "GAME_OVER",
              "DUPLICATE_GUESS",
              "PLAN_NOT_FOUND",
              "PLAN_NOT_PURCHASABLE",
              "PAYMENTS_UNAVAILABLE",
              "PAYMENT_NOT_FOUND",
              "INVALID_WEBHOOK_SIGNATURE",
              "VENUE_NOT_FOUND",
              "NOT_VENUE_OWNER"
            ],
            "description": "Stable error code: switch on it and show your own (translated) text",
            "example": "AGE_RESTRICTED"
          },
          "message": {
            "type": "string",
            "description": "Human hint in English; may change, do not parse it",
            "example": "This drink contains alcohol: sign in with a verified adult account to see it"
          },
          "error": {
            "type": "string",
            "example": "ForbiddenError"
          },
          "details": {
            "description": "Only with VALIDATION_FAILED: one entry per invalid field",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "requestId": {
            "description": "Same as the X-Request-Id header: quote it when reporting a problem",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "error"
        ],
        "additionalProperties": false,
        "description": "Every error response has this shape."
      },
...
      "ImageSizes": {
        "nullable": true,
        "description": "The drink picture in three sizes",
        "type": "object",
        "properties": {
          "small": {
            "type": "string"
          },
          "medium": {
            "type": "string"
          },
          "large": {
            "type": "string"
          }
        },
        "required": [
          "small",
          "medium",
          "large"
        ],
        "additionalProperties": false
      },
      "Ingredient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tequila"
          },
          "nameEs": {
            "nullable": true,
            "example": "Tequila",
            "description": "Spanish name, if known",
            "type": "string"
          },
          "measure": {
            "nullable": true,
            "example": "1 1/2 oz",
            "type": "string"
          },
          "image": {
            "nullable": true,
            "description": "Small picture of the ingredient",
            "type": "string"
          }
        },
        "required": [
          "name",
          "nameEs",
          "measure",
          "image"
        ],
        "additionalProperties": false
      },
...
      "Drink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "11007"
          },
          "name": {
            "type": "string",
            "example": "Margarita"
          },
          "category": {
            "nullable": true,
            "example": "Ordinary Drink",
            "type": "string"
          },
          "categoryEs": {
            "nullable": true,
            "example": "Trago clásico",
            "type": "string"
          },
          "alcoholic": {
            "type": "boolean"
          },
          "glass": {
            "nullable": true,
            "example": "Cocktail glass",
            "type": "string"
          },
          "glassEs": {
            "nullable": true,
            "example": "Copa de cóctel",
            "type": "string"
          },
          "iba": {
            "nullable": true,
            "description": "IBA official list, if any",
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "nullable": true,
            "type": "string"
          },
          "images": {
            "$ref": "#/components/schemas/ImageSizes"
          },
          "video": {
            "nullable": true,
            "type": "string"
          },
          "instructions": {
            "type": "object",
            "properties": {
              "en": {
                "nullable": true,
                "type": "string"
              },
              "es": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "en",
              "es"
            ],
            "additionalProperties": false
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            }
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "categoryEs",
          "alcoholic",
          "glass",
          "glassEs",
          "iba",
          "tags",
          "image",
          "images",
          "video",
          "instructions",
          "ingredients"
        ],
        "additionalProperties": false
      },
...
      "DrinkSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "11007"
          },
          "name": {
            "type": "string",
            "example": "Margarita"
          },
          "image": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "image"
        ],
        "additionalProperties": false
      },
...
      "Mood": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "hungover"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "guayabo",
              "resaca"
            ]
          },
          "description": {
            "type": "string"
          },
          "alcoholic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string",
                "enum": [
                  "any"
                ]
              }
            ],
            "description": "Whether the mood forces alcoholic/alcohol-free drinks"
          }
        },
        "required": [
          "key",
          "aliases",
          "description",
          "alcoholic"
        ],
        "additionalProperties": false
      },
      "MoodRecommendation": {
        "type": "object",
        "properties": {
          "mood": {
            "type": "string"
          },
          "why": {
            "type": "string"
          },
          "drink": {
            "$ref": "#/components/schemas/Drink"
          },
          "alternatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DrinkSummary"
            }
          }
        },
        "required": [
          "mood",
          "why",
          "drink",
          "alternatives"
        ],
        "additionalProperties": false
      },
      "PantrySuggestions": {
        "type": "object",
        "properties": {
          "have": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What was understood, synonyms included"
          },
          "canMake": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Drink"
            }
          },
          "almost": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "11007"
                },
                "name": {
                  "type": "string",
                  "example": "Margarita"
                },
                "image": {
                  "nullable": true,
                  "type": "string"
                },
                "missing": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "id",
                "name",
                "image",
                "missing"
              ],
              "additionalProperties": false
            }
          },
          "shoppingTips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingTip"
            }
          },
          "totals": {
            "type": "object",
            "properties": {
              "canMake": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "almost": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "canMake",
              "almost"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "have",
          "canMake",
          "almost",
          "shoppingTips",
          "totals"
        ],
        "additionalProperties": false
      },
      "ShoppingTip": {
        "type": "object",
        "properties": {
          "buy": {
            "type": "string",
            "example": "gin"
          },
          "unlocks": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "drinks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "buy",
          "unlocks",
          "drinks"
        ],
        "additionalProperties": false
      },
...
    }
  }
}